#5 - code modularizing Make each module give a value that it is supposed to do. Separation of concerns
